adding tests and modifyning code for dwa to be equals if they have the same Nans
